chat-with-documents-local-llm/pyproject.toml
[tool.poetry]
name = "chat-with-docs-saas"
version = "0.1.1"
description = "Integrate Celery with Redis, Enhance Logging, UX, and Task Feedback"
authors = ["Mithun Parab"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"

# LangChain & AI
langchain = "^0.2.1"
langchain-community = "^0.2.1"
langchain-groq = "^0.1.5"
langchain-google-genai = "^1.0.5"
langchain-ollama = "^0.1.0"
langchain-chroma = "^0.1.1"
groq = "^0.9.0"

# Document Loaders
unstructured = "^0.14.9"
beautifulsoup4 = "^4.12.3"
pypdf = "^4.2.0"
python-docx = "^1.1.2"

# Vector Search
chromadb = "^0.5.3"

# Database (Postgres)
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
rank-bm25 = "^0.2.2"

# Authentication
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx-oauth = "*" # Let poetry choose the best version

# Storage (S3/MinIO)
boto3 = "^1.34.111"

# For frontend/CLI interaction
requests = "^2.32.3"

# Background Tasks & Streaming
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.0.4"
sse-starlette = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.main.api"